extract FlussDataTypeToPaimonDataType to a common class
